fix search result slice bug
